CI: Use PyCall-managed Conda python
On some CI platforms (e.g., Ubuntu), the system `python` is
picked up by PyCall and therefore Obspy is not installed and
not available.  Set the environment variable `PYTHON=""` for
the package installation action to get PyCall to use its own
Conda-managed `python`, which will automatically install Obspy
for us.